implemented group related methods | renamed some files | modified controllers social.wall & group.chat

+++ app/controllers/user/group/remove.user.group.js
@@ -4,6 +4,15 @@ const GroupMembers = db2.members;
 
 module.exports.removeUserFromGroup = async(group_id,reference_number) => {
   try {
+
+    const count = await GroupMembers.count({
+      where: { group_id, reference_number }
+    });
+
+    if (count === 0) {
+      return [false, 'No group membership found to remove'];
+    }
+	  
     await GroupMembers.destroy({
       where: { group_id, reference_number },
     });
